refactor(services): extract useLastDeployedCommit hook
